Remove non-singleton appclient instance
We had two appclients running/refreshing in the scheduler, this makes both
call sites use the singleton one

[#142730507]
